Fix delete for like, reply and bookmark, add find a tweet by id, fix create reply, fix query for trending tweets, refactor tweet controller and add UnauthorizedException for interacting with likes and bookmarks.

+++ backend/src/main/java/com/twitterclone/backend/model/services/TweetServiceJpa.java
@@ -1,23 +1,17 @@
 package com.twitterclone.backend.model.services;
 
-import com.twitterclone.backend.dto.DisplayTweetDto;
 import com.twitterclone.backend.model.DisplayTweet;
-import com.twitterclone.backend.model.entities.Hashtag;
-import com.twitterclone.backend.model.entities.LikeTweet;
-import com.twitterclone.backend.model.entities.Tweet;
-import com.twitterclone.backend.model.entities.User;
+import com.twitterclone.backend.model.entities.*;
 import com.twitterclone.backend.model.exceptions.EntityNotFoundException;
 import com.twitterclone.backend.model.exceptions.ReactionAlreadyExistsException;
+import com.twitterclone.backend.model.exceptions.UnauthorizedException;
 import com.twitterclone.backend.model.repositories.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class TweetServiceJpa implements TweetService {
@@ -38,6 +32,23 @@ public class TweetServiceJpa implements TweetService {
         this.bookmarkRepo = bookmarkRepo;
     }
 
+    private void updateTweetDetails(DisplayTweet tweetDetails, long userId) {
+        long tweetId = tweetDetails.getTweet().getId();
+
+        tweetDetails.setLikeCount(likeTweetRepo.countLikesByTweetId(tweetId));
+        tweetDetails.setReplyCount(replyRepo.countReplyByTweetId(tweetId));
+        tweetDetails.setBookmarkCount(bookmarkRepo.countBookmarkByTweetId(tweetId));
+        tweetDetails.setReplied(!replyRepo.findReplyByUserIdAndTweetId(userId, tweetId).isEmpty());
+
+        Optional<LikeTweet> oLike = likeTweetRepo.findLikeByUserIdAndTweetId(userId, tweetId);
+        tweetDetails.setLiked(oLike.isPresent());
+        oLike.ifPresent(l -> tweetDetails.setLikeId(l.getId()));
+
+        Optional<Bookmark> oBookmark = bookmarkRepo.findBookmarkByUserIdAndTweetId(userId, tweetId);
+        tweetDetails.setBookmarked(oBookmark.isPresent());
+        oBookmark.ifPresent(b -> tweetDetails.setBookmarkId(b.getId()));
+    }
+
     public Tweet createTweet(Tweet tweet, long userId) throws EntityNotFoundException {
         User user = userRepo.findById(userId)
                 .orElseThrow(() -> new EntityNotFoundException("Entity not found", User.class.getName()));
@@ -49,18 +60,19 @@ public class TweetServiceJpa implements TweetService {
     }
 
     @Override
+    public DisplayTweet findTweetById(long tweetId) throws EntityNotFoundException {
+        return tweetRepo.findById(tweetId)
+                .map(DisplayTweet::new)
+                .orElseThrow(() -> new EntityNotFoundException("Entity not found", Tweet.class.getName()));
+    }
+
+    @Override
     public Page<DisplayTweet> getTrendingTweets(Pageable pageable, long userId) {
         Page<DisplayTweet> tweets = tweetRepo
                 .getTweetsByLikesAndCommentsDesc(pageable)
                 .map(DisplayTweet::new);
 
-        tweets.forEach(t -> {
-            t.setLikeCount(likeTweetRepo.countLikesByTweetId(t.getTweet().getId()));
-            t.setReplyCount(replyRepo.countReplyByTweetId(t.getTweet().getId()));
-            t.setBookmarkCount(bookmarkRepo.countBookmarkByTweetId(t.getTweet().getId()));
-            t.setLiked(likeTweetRepo.findLikeByUserIdAndTweetId(userId, t.getTweet().getId()).isPresent());
-            t.setBookmarked(bookmarkRepo.findBookmarkByUserIdAndTweetId(userId, t.getTweet().getId()).isPresent());
-        });
+        tweets.forEach(t -> updateTweetDetails(t, userId));
 
         return tweets;
     }
@@ -72,7 +84,7 @@ public class TweetServiceJpa implements TweetService {
         Tweet tweet = tweetRepo.findById(tweetId)
                 .orElseThrow(() -> new EntityNotFoundException("Entity not found", Tweet.class.getName()));
 
-        // If the user has already reacted to the tweet
+        // Check if the user already liked the tweet
         if (likeTweetRepo.findLikeByUserIdAndTweetId(userId, tweetId).isPresent()) {
             throw new ReactionAlreadyExistsException("Reaction already present", LikeTweet.class.getName());
         }
@@ -82,39 +94,75 @@ public class TweetServiceJpa implements TweetService {
 
         likeTweetRepo.save(like);
 
-        // To update the like counts of the tweet
+        // Update tweet details after like
         DisplayTweet tweetDetails = new DisplayTweet(tweet);
-        tweetDetails.setLikeCount(likeTweetRepo.countLikesByTweetId(tweetId));
-        tweetDetails.setReplyCount(replyRepo.countReplyByTweetId(tweetId));
-        tweetDetails.setBookmarkCount(bookmarkRepo.countBookmarkByTweetId(tweetId));
-        tweetDetails.setLiked(likeTweetRepo.findLikeByUserIdAndTweetId(userId, tweetId).isPresent());
-        tweetDetails.setBookmarked(bookmarkRepo.findBookmarkByUserIdAndTweetId(userId, tweetId).isPresent());
+        updateTweetDetails(tweetDetails, userId);
 
         return tweetDetails;
     }
 
-    public DisplayTweet deleteLikeFromTweet(long tweetId, long userId) throws EntityNotFoundException {
-        // if the user, tweet or the like are not in the database throw an exception
-        userRepo.findById(userId)
-                .orElseThrow(() -> new EntityNotFoundException("Entity not found", User.class.getName()));
-
-        DisplayTweet tweet = tweetRepo.findById(tweetId)
-                .map(DisplayTweet::new)
-                .orElseThrow(() -> new EntityNotFoundException("Entity not found", Tweet.class.getName()));
-
-        LikeTweet like = likeTweetRepo.findLikeByUserIdAndTweetId(userId, tweetId)
+    @Override
+    public DisplayTweet deleteLikeFromTweet(long id, long userId) throws EntityNotFoundException, UnauthorizedException {
+        LikeTweet like = likeTweetRepo.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Entity not found", LikeTweet.class.getName()));
 
-        // delete
+        if (like.getUser().getId() != userId) {
+            throw new UnauthorizedException("User is not authorized to delete this", LikeTweet.class.getName());
+        }
+
+        DisplayTweet tweet = new DisplayTweet(like.getTweet());
+        long tweetId = tweet.getTweet().getId();
+
         likeTweetRepo.deleteById(like.getId());
 
-        // Return the tweet's details
-        tweet.setLikeCount(likeTweetRepo.countLikesByTweetId(tweetId));
-        tweet.setReplyCount(replyRepo.countReplyByTweetId(tweetId));
-        tweet.setBookmarkCount(bookmarkRepo.countBookmarkByTweetId(tweetId));
-        tweet.setLiked(likeTweetRepo.findLikeByUserIdAndTweetId(userId, tweetId).isPresent());
-        tweet.setBookmarked(bookmarkRepo.findBookmarkByUserIdAndTweetId(userId, tweetId).isPresent());
+        // Update tweet details after like deletion
+        updateTweetDetails(tweet, userId);
 
         return tweet;
     }
+
+    @Override
+    public DisplayTweet createBookmarkToTweet(Bookmark bookmark, long userId, long tweetId) throws EntityNotFoundException, ReactionAlreadyExistsException {
+        User user = userRepo.findById(userId)
+                .orElseThrow(() -> new EntityNotFoundException("Entity not found", User.class.getName()));
+        Tweet tweet = tweetRepo.findById(tweetId)
+                .orElseThrow(() -> new EntityNotFoundException("Entity not found", Tweet.class.getName()));
+
+        // Check if the user already bookmarked the tweet
+        if (bookmarkRepo.findBookmarkByUserIdAndTweetId(userId, tweetId).isPresent()) {
+            throw new ReactionAlreadyExistsException("Reaction already present", Bookmark.class.getName());
+        }
+
+        bookmark.setUser(user);
+        bookmark.setTweet(tweet);
+
+        bookmarkRepo.save(bookmark);
+
+        // Update tweet details after bookmark
+        DisplayTweet tweetDetails = new DisplayTweet(tweet);
+        updateTweetDetails(tweetDetails, userId);
+
+        return tweetDetails;
+    }
+
+    @Override
+    public DisplayTweet deleteBookmarkFromTweet(long id, long userId) throws EntityNotFoundException, UnauthorizedException {
+        Bookmark bookmark = bookmarkRepo.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Entity not found", Bookmark.class.getName()));
+
+        if (bookmark.getUser().getId() != userId) {
+            throw new UnauthorizedException("User is not authorized to delete this", Bookmark.class.getName());
+        }
+
+        DisplayTweet tweet = new DisplayTweet(bookmark.getTweet());
+        long tweetId = tweet.getTweet().getId();
+
+        bookmarkRepo.deleteById(bookmark.getId());
+
+        // Update tweet details after bookmark deletion
+        updateTweetDetails(tweet, userId);
+
+        return tweet;
+    }
+
 }
